convert to 3 dot menu
